Chakra Edits
Fixed chakra being used while in Creative mode
Added additional check to player chakra v jutsu chakra usage to 
ensure it doesn't go negative

+++ src/main/java/com/cdc/naruto/capabilities/chakra/ICapChakra.java
@@ -3,9 +3,9 @@ package com.cdc.naruto.capabilities.chakra;
 import com.cdc.naruto.capabilities.ICapability;
 
 public interface ICapChakra extends ICapability {
-	void diminishChakra(float amount);
+	void diminishChakra(int amount);
 	void replenishChakra();
-	void fillChakra(float amount);
-	void setChakra(float amount);
-	float getChakra();
+	void fillChakra(int amount);
+	void setChakra(int amount);
+	int getChakra();
 }
